requirements.txt
firebase_admin
google-cloud-firestore
streamlit~=0.86.0
